+++ src/aslm/view/main_window_content/stage_control/images/greyup.png
